#475 moved Theme and ThemeManager classes to a new file ThemeManager.cs
